Rename FaultTolerantExecutionTest to FaultTolerantExecutionCoverage
This avoids Test[...]Test names.

I used the following script.
```
set -euo pipefail

files=()
while IFS= read -r f; do
  files+=("$f")
done < <(git ls-files '*FaultTolerantExecutionTest.java')

names=()
for f in "${files[@]}"; do
  base=$(basename "$f" .java)
  names+=("$base")
  new="${f%FaultTolerantExecutionTest.java}"
  new="${new}FaultTolerantExecutionCoverage.java"
  git mv "$f" "$new"
done

for n in "${names[@]}"; do
  new="${n%FaultTolerantExecutionTest}FaultTolerantExecutionCoverage"
  git grep -lz "$n" | while IFS= read -r -d '' file; do
    perl -i -pe "s/\b\Q${n}\E\b/${new}/g" "$file"
  done
done

```
